+++ R/read_github_file.R
@@ -1,12 +1,10 @@
 #' Read a file from a GitHub CRAN repository
 #' @description Read a file from a GitHub CRAN repository.
-#' @param pac `character` package name.
-#' @param version `character` package version.
-#' @param file `character` file name to read. Possible values are `DESCRIPTION` and `NAMESPACE`.
-#' @param repos `character` vector repositories URLs to use. Used only for the validation. Default `https://cran.rstudio.com/`
+#' @inheritParams standard_args
 #' @note if the file is not found in the GitHub repository, it will try to find it in the CRAN archive.
 #' @keywords internal
 read_github_file <- function(pac, version, file, repos = "https://cran.rstudio.com/") {
+  stopifnot(is_online())
   ee <- tempfile()
   d_url <- sprintf(
     "https://raw.githubusercontent.com/cran/%s/%s/%s",
@@ -24,9 +22,10 @@ read_github_file <- function(pac, version, file, repos = "https://cran.rstudio.c
     silent = TRUE
   )
   if (inherits(tt, "try-error")) {
-    result <- cran_archive_file(pac, version, repos, file)
+    result <- read_cran_file(pac, version, file, repos)
   } else {
-    res <- readLines(ee)
+    result <- readLines(ee)
     unlink(ee)
   }
+  result
 }